Move token counting to model detail.
